renamed project to StrainedBandstructures, renamed MaterialDataset to MaterialParameters and MaterialData to HetereoStructureData (now in its own file); now seperate tensor supertypes for elasticity and piezoeletricity, SparseMatrixTensor and DenseMatrixTensor replaces the Custom***Tensors, all tensors implement apply_tensor!, adjusted documentation
