+++ docs/make.jl
@@ -1,10 +1,10 @@
 using Documenter
-using NanoWiresJulia
+using StrainedBandstructures
 
 push!(LOAD_PATH,"../src/")
 
 makedocs(
-		modules = [NanoWiresJulia],
+		modules = [StrainedBandstructures],
 		sitename = "StrainedBandStructures.jl",
 		authors = "Patricio Farrell, Yiannis Hadjimichael, Christian Merdon",
 		repo = "github.com/hadjimy/StrainedBandstructures",
